+++ src/Main.kt
@@ -1,7 +1,25 @@
-import com.nativebuild.Build
-import com.nativebuild.Menu
-import com.nativebuild.Info
-import com.nativebuild.License
+/*
+   Copyright 2020 Joshua Kent
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
+
+package com.nativebuild
+
+import com.nativebuild.util.Build
+import com.nativebuild.util.Menu
+import com.nativebuild.util.Info
+import com.nativebuild.util.License
 
 
 fun main(args: Array<String>) {
@@ -25,7 +43,7 @@ fun main(args: Array<String>) {
             in arrayOf("y", "yes") -> run = true
             in arrayOf("n", "no") -> System.exit(0)
             "info" -> System.exit(0) // TODO
-            "license" -> System.exit(0) // TODO
+            "license" -> License.displayLicense()
             else -> System.exit(0) // TODO
         }
     } else {
@@ -38,7 +56,7 @@ fun main(args: Array<String>) {
             in arrayOf("y", "yes") -> run = true
             in arrayOf("n", "no") -> System.exit(0)
             "info" -> System.exit(0) // TODO
-            "license" -> System.exit(0) // TODO
+            "license" -> License.displayLicense()
             else -> System.exit(0) // TODO
         }