+++ src/com/nativebuild/License.kt
@@ -1,8 +1,0 @@
-package com.nativebuild
-
-object License {
-    fun displayLicense() {
-        // TODO
-        println(com.nativebuild.LICENSE)
-    }
-}